Improve testability of the code
