+++ lib/cast_vote_msg.js
@@ -2,7 +2,7 @@ const Assert = require('assert-plus')
 const Vote = require('./vote_entity')
 const Shapes = require('./shapes')
 const { fetchProp } = require('./utils')
-const { ValidationError, NotFoundError } = require('./errors')
+const { NotFoundError } = require('./errors')
 const VoteStats = require('./vote_stats')
 const PollRating = require('./poll_rating')
 const Reply = require('./reply')
@@ -51,23 +51,7 @@ module.exports = function (opts = {}) {
         data: { poll_stats }
       }))
     } catch (err) {
-      if (err instanceof ValidationError) {
-        return reply(null, Reply.invalidFieldOfValidationError(err))
-      }
-
-      if (err instanceof NotFoundError) {
-        const details = (() => {
-          if (typeof err.message === 'string') {
-            return { details: { what: err.message } }
-          }
-
-          return {}
-        })()
-
-        return reply(null, Reply.notFound(details))
-      }
-
-      return reply(err)
+      return Reply.handleAppError(err, reply)
     }
   }
