+++ lib/get_poll_msg.js
@@ -1,7 +1,6 @@
 const Assert = require('assert-plus')
 const Shapes = require('./shapes')
 const { fetchProp } = require('./utils')
-const { ValidationError } = require('./errors')
 const Reply = require('./reply')
 
 module.exports = function (opts = {}) {
@@ -26,11 +25,7 @@ module.exports = function (opts = {}) {
         data: { poll: poll.data$(false) }
       }))
     } catch (err) {
-      if (err instanceof ValidationError) {
-        return reply(null, Reply.invalidFieldOfValidationError(err))
-      }
-
-      return reply(err)
+      return Reply.handleAppError(err, reply)
     }
   }
 }
