+++ lib/open_poll_msg.js
@@ -2,7 +2,6 @@ const Assert = require('assert-plus')
 const Joi = require('joi')
 const Shapes = require('./shapes')
 const { fetchProp } = require('./utils')
-const { ValidationError } = require('./errors')
 const Vote = require('./vote_entity')
 const Reply = require('./reply')
 
@@ -24,11 +23,7 @@ module.exports = function (opts = {}) {
         data: { poll: opened_poll.data$(false) }
       }))
     } catch (err) {
-      if (err instanceof ValidationError) {
-        return reply(null, Reply.invalidFieldOfValidationError(err))
-      }
-
-      return reply(err)
+      return Reply.handleAppError(err, reply)
     }
   }
