+++ lib/reply.js
@@ -84,6 +84,42 @@ class Reply {
 
     return shape
   }
+
+  static handleAppError(err, reply) {
+    //
+    // NOTE: This function is meant to handle general app errors, e.g.
+    // related to an entity not being found or failed validation.
+    // If you have any action-specific errors that you want to handle,
+    // please handle them in the corresponding message handler, e.g.
+    // ```
+    //   try { ... }
+    //   catch (err) {
+    //     if (err instanceof MyPreciousError) {
+    //       return reply(null, ...)
+    //     }
+    //
+    //     return handleAppError(err, reply)
+    //   }
+    // ```
+    //
+    if (err instanceof ValidationError) {
+      return reply(null, Reply.invalidFieldOfValidationError(err))
+    }
+
+    if (err instanceof NotFoundError) {
+      const details = (() => {
+        if (typeof err.message === 'string') {
+          return { details: { what: err.message } }
+        }
+
+        return {}
+      })()
+
+      return reply(null, Reply.notFound(details))
+    }
+
+    return reply(err)
+  }
 }
 
 module.exports = Reply
